find donor table view drag, few other changes

+++ bloodBank/bloodBank/UITableViewZoomController/TTUITableViewZoomController.m
@@ -57,8 +57,8 @@
                          } completion:nil];
         
         
-        currentMaxDisplayedCell = indexPath.row;
-        currentMaxDisplayedSection = indexPath.section;
+        currentMaxDisplayedCell = (int)indexPath.row;
+        currentMaxDisplayedSection = (int)indexPath.section;
     }
 }
